Remove SyncQueue and migrate more ES6 classes (#2674)
* Remove SyncQueue and migrates more ES6 classes

* More SortedLoopArray to core and remove from pc namespace

* Re-adding SortedLoopArray for now to pass unit tests
