+++ src/index.js
@@ -19,10 +19,10 @@ export { string } from './core/string.js';
 export { Color } from './core/color.js';
 export { EventHandler } from './core/event-handler.js';
 export { IndexedList } from './core/indexed-list.js';
+export { SortedLoopArray } from './core/sorted-loop-array.js';
 export { Tags } from './core/tags.js';
 export { Timer, now } from './core/time.js';
 export { URI, createURI } from './core/uri.js';
-export { RefCountedObject } from './core/ref-counted-object.js';
 
 // NET
 export { http, Http } from './net/http.js';
@@ -248,7 +248,8 @@ export { PostEffectQueue } from './framework/components/camera/post-effect-queue
 export * from './framework/components/rigid-body/constants.js';
 export { RigidBodyComponent } from './framework/components/rigid-body/component.js';
 export { RigidBodyComponentSystem, ContactPoint, ContactResult, RaycastResult, SingleContactResult } from './framework/components/rigid-body/system.js';
-export { SceneRegistry, SceneRegistryItem } from './framework/scene-registry.js';
+export { SceneRegistry } from './framework/scene-registry.js';
+export { SceneRegistryItem } from './framework/scene-registry-item.js';
 export * from './framework/components/screen/constants.js';
 export { ScreenComponent } from './framework/components/screen/component.js';
 export { ScreenComponentSystem } from './framework/components/screen/system.js';
@@ -261,7 +262,6 @@ export { ScrollbarComponentSystem } from './framework/components/scrollbar/syste
 export * from './framework/components/scroll-view/constants.js';
 export { ScrollViewComponent } from './framework/components/scroll-view/component.js';
 export { ScrollViewComponentSystem } from './framework/components/scroll-view/system.js';
-export { SortedLoopArray } from './framework/utils/sorted-loop-array.js';
 export { SoundSlot } from './framework/components/sound/slot.js';
 export { SoundComponent } from './framework/components/sound/component.js';
 export { SoundComponentSystem } from './framework/components/sound/system.js';
